IGのメディアアップロード状況確認を共通化
