+++ src/apis/post_ig_video.rs
@@ -1,3 +1,4 @@
+use crate::apis::check_ig_media::check_ig_media_loop;
 use crate::*;
 
 impl Fbapi {
@@ -24,7 +25,7 @@ impl Fbapi {
         )
         .await?;
 
-        check_loop(
+        check_ig_media_loop(
             &self.make_path(&format!(
                 "{}?fields=status_code&access_token={}",
                 creation_id, access_token
@@ -73,7 +74,7 @@ impl Fbapi {
         )
         .await?;
 
-        check_loop(
+        check_ig_media_loop(
             &self.make_path(&format!(
                 "{}?fields=status_code&access_token={}",
                 container_id, access_token
@@ -129,45 +130,6 @@ async fn post(
         Some(s) => Ok(s.to_owned()),
         None => return Err(FbapiError::UnExpected(res)),
     }
-}
-
-async fn check(
-    path: &str,
-    retry_count: usize,
-    client: &reqwest::Client,
-    log: &impl Fn(LogParams),
-) -> Result<String, FbapiError> {
-    let log_params = LogParams::new(&path, &vec![]);
-    let res = execute_retry(
-        retry_count,
-        || async { client.get(path).send().await.map_err(|e| e.into()) },
-        log,
-        log_params,
-    )
-    .await?;
-    match res["status_code"].as_str() {
-        Some(s) => Ok(s.to_owned()),
-        None => return Err(FbapiError::UnExpected(res)),
-    }
-}
-
-async fn check_loop(
-    path: &str,
-    check_retry_count: usize,
-    check_video_delay: usize,
-    retry_count: usize,
-    client: &reqwest::Client,
-    log: &impl Fn(LogParams),
-) -> Result<(), FbapiError> {
-    for _ in 0..check_retry_count {
-        match check(path, retry_count, client, log).await?.as_str() {
-            "FINISHED" => return Ok(()),
-            "IN_PROGRESS" => {}
-            _ => return Err(FbapiError::VideoError),
-        }
-        sleep(check_video_delay).await;
-    }
-    Err(FbapiError::VideoTimeout)
 }
 
 async fn publish(
